refactor: export the building blocks for building a controller on the v2 stack
Nothing outside this module can currently assemble a controller on kagent's
own v2 components. api/database exports the Client interface but not a
constructor, and grpcserver is internal, so a downstream can only ever
receive a Client from core/pkg/app -- which never sets A2AHandler and so
never registers the A2A service.

Two changes, no behaviour change:

  * core/pkg/database re-exports Connect/NewClient/ResolveURL and aliases
    PostgresConfig. The implementation stays in core/internal/database.
  * core/internal/grpcserver moves to core/pkg/grpcserver.

Config keeps its fields typed on internal service packages. That is fine for
external callers: omitting a struct field never requires importing its type,
and the ones a downstream cannot name are all optional -- SystemService and
MethodPolicies self-default, so DefaultMethodPolicies still applies.
