+++ go/core/pkg/grpcserver/prompttemplate.go
