+++ go/core/pkg/grpcserver/prompttemplate_test.go
